run useeffect in ManuscriptCard instead
